third_party/WebKit/LayoutTests/shadow-dom/css-focus-pseudo-match-shadow-host2-expected.txt

(1/6) Testing how :focus matches on shadow host with delegatesFocus=false.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
(2/6) Testing how :focus matches on shadow host with tabindex=-1, delegatesFocus=false.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
(3/6) Testing how :focus matches on shadow host with tabindex=0, delegatesFocus=false.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
(4/6) Testing how :focus matches on shadow host with delegatesFocus=true.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
(5/6) Testing how :focus matches on shadow host with tabindex=-1, delegatesFocus=true.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
(6/6) Testing how :focus matches on shadow host with tabindex=0, delegatesFocus=true.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
